mal_fun/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
mal-api.py==0.10.1
numpy==2.2.6
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
